Refactor game logic to Kotlin and remove Java files
- Migrated `Combo.java` to `Combo.kt`.
- Migrated `GameState.java` to `GameState.kt`.
- Removed `EvaluationThread.java` as its functionality is now integrated within `GameState.kt`.
- Updated `GamePlayNew.java` to use the new Kotlin classes `Combo` and `GameState`.
- Adjusted `LifeBar.kt` to accept `Short` for `typeTap` in `updateLife` function, aligning with changes in `Combo.kt`.

+++ app/src/main/java/com/kyagamy/step/game/newplayer/LifeBar.kt
@@ -116,7 +116,7 @@ class LifeBar(
         }
     }
 
-    fun updateLife(typeTap: Byte,combo: Int) {
+    fun updateLife(typeTap: Short, combo: Int) {
         when (typeTap) {
             Combo.VALUE_PERFECT, Combo.VALUE_GREAT -> life+=1*abs(combo)
             Combo.VALUE_BAD -> life-=0.3f*abs(combo)
